Initial work for v3.0.0
