+++ src/language/providers/problemProvider.ts
@@ -4,9 +4,8 @@ import {
 } from "../../database/schemas";
 import { remoteAssistIsEnabled } from "./logic/available";
 import Configuration from "../../configuration";
-import { SQLStatementChecker, SqlSyntaxError } from "../../connection/syntaxChecker";
+import { ValidateStatementComponent, SqlSyntaxError, MAX_STATEMENT_COUNT, VALID_STATEMENT_LENGTH } from "../../connection/components/validateStatement";
 import { StatementGroup, StatementType } from "../sql/types";
-import { MAX_STATEMENT_COUNT, VALID_STATEMENT_LENGTH } from "../../connection/syntaxChecker/checker";
 import { getSqlDocument, isSafeDocument } from "./logic/parse";
 import path from "path";
 
@@ -51,7 +50,7 @@ function shouldShowWarnings() {
 const CHECKER_AVAILABLE_CONTEXT = `vscode-db2i.syntax.checkerAvailable`;
 const CHECKER_RUNNING_CONTEXT = `vscode-db2i.syntax.checkerRunning`;
 
-const checkerAvailable = () => (SQLStatementChecker.get() !== undefined);
+const checkerAvailable = () => (ValidateStatementComponent.get() !== undefined);
 
 export function setCheckerAvailableContext(additionalState = true) {
   const available = checkerAvailable() && additionalState;
@@ -119,7 +118,7 @@ interface SqlDiagnostic extends Diagnostic {
 }
 
 async function validateSqlDocument(document: TextDocument, specificStatement?: number) {
-  const checker = SQLStatementChecker.get();
+  const checker = ValidateStatementComponent.get();
   const job = remoteAssistIsEnabled(true);
   if (checker && job) {
     const basename = document.fileName ? path.basename(document.fileName) : `Untitled`;
