+++ src/language/providers/statusProvider.ts
@@ -1,7 +1,7 @@
 import { Disposable, languages, LanguageStatusSeverity } from "vscode";
-import { SQLStatementChecker } from "../../connection/syntaxChecker";
 import { getCheckerTimeout } from "./problemProvider";
 import { useSystemNames } from "./logic/available";
+import { ValidateStatementComponent } from "../../connection/components/validateStatement";
 
 export class Db2StatusProvider extends Disposable {
   private item = languages.createLanguageStatusItem(`sql`, {language: `sql`});
@@ -14,7 +14,7 @@ export class Db2StatusProvider extends Disposable {
 
   setState(hasJob: Boolean) {
     if (hasJob) {
-      const checker = SQLStatementChecker.get();
+      const checker = ValidateStatementComponent.get();
       const checkerTimeout = getCheckerTimeout() / 1000;
       this.item.text = `SQL assistance available. ${checker ? `Syntax checking enabled (every ${checkerTimeout}s after editing)` : `Syntax checking not available.`}`;
       this.item.detail = `You're connected to an IBM i. ${checker ? `You can use the advanced SQL language tooling.` : `Syntax checking not available. This means that the syntax checker did not install when connecting to this system.`}`;
